se der clone de uma olhada no README

+++ app/Http/Controllers/CursoController.php
@@ -5,62 +5,71 @@ namespace App\Http\Controllers;
 use App\Models\Curso;
 use Illuminate\Http\Request;
 
-class CursoController extends Controller {
+class CursoController extends Controller
+{
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->authorizeResource(Curso::class, 'curso');
     }
 
-    public function index() {
+    public function index()
+    {
 
-        // $this->authorize('viewAny',  Curso::class);
+        $this->authorize('viewAny',  Curso::class);
+        
         $cursos = Curso::all();
         return view('cursos.index', compact('cursos'));
     }
 
-    public function create() {
+    public function create()
+    {
 
-        // $this->authorize('create',  Curso::class);
+        $this->authorize('create',  Curso::class);
         return view('cursos.create');
     }
 
-    public function store(Request $request) {
-        // $this->authorize('create',  Curso::class);
+    public function store(Request $request)
+    {
+        $this->authorize('create',  Curso::class);
 
         $obj = new Curso();
-        $obj->nome = mb_strtoupper($request->nome, 'UTF-8');   
+        $obj->nome = mb_strtoupper($request->nome, 'UTF-8');
         $obj->save();
 
         return redirect()->route('cursos.index');
     }
 
-    public function show(Curso $curso) {
-        
-        // $this->authorize('view', $curso);
+    public function show(Curso $curso)
+    {
 
-        if(isset($curso)) {
+        $this->authorize('view', $curso);
+
+        if (isset($curso)) {
             return view('cursos.show', compact('curso'));
         }
 
-        return "<h1>Curso não Encontrado!</h1>";        
+        return "<h1>Curso não Encontrado!</h1>";
     }
 
-    public function edit(Curso $curso) {
-        
-        // $this->authorize('update', $curso);
+    public function edit(Curso $curso)
+    {
 
-        if(isset($curso)) {
+        $this->authorize('update', $curso);
+
+        if (isset($curso)) {
             return view('cursos.edit', compact('curso'));
         }
 
-        return "<h1>Curso não Encontrado!</h1>";        
+        return "<h1>Curso não Encontrado!</h1>";
     }
 
-    public function update(Request $request, Curso $curso) {
-        // $this->authorize('update', $curso);
+    public function update(Request $request, Curso $curso)
+    {
+        $this->authorize('update', $curso);
 
-        if(isset($curso)) {
-            $curso->nome = mb_strtoupper($request->nome, 'UTF-8');   
+        if (isset($curso)) {
+            $curso->nome = mb_strtoupper($request->nome, 'UTF-8');
             $curso->save();
             return redirect()->route('cursos.index');
         }
@@ -68,11 +77,12 @@ class CursoController extends Controller {
         return "<h1>Curso não Encontrado!</h1>";
     }
 
-    public function destroy(Curso $curso) {
-        
-        // $this->authorize('delete', $curso);
+    public function destroy(Curso $curso)
+    {
 
-        if(isset($curso)) {
+        $this->authorize('delete', $curso);
+
+        if (isset($curso)) {
             $curso->delete();
             return redirect()->route('cursos.index');
         }
